chore: move out capabilities
